chromium/third_party/blink/web_tests/virtual/no-auto-wpt-origin-isolation/external/wpt/origin-isolation/getter-data-url.https-expected.txt

This is a testharness.js-based test.
FAIL originIsolationRestricted must equal false assert_equals: expected (boolean) false but got (undefined) undefined
Harness: the test ran to completion.
